+++ packages/core-cairo/src/index.ts
@@ -20,6 +20,6 @@ export { OptionsError } from './error';
 export type { Kind } from './kind';
 export { sanitizeKind } from './kind';
 
-export { contractsVersion } from './utils/version';
+export { contractsVersion, contractsVersionTag } from './utils/version';
 
 export { erc20, erc721, utils } from './api';