+++ packages/core-cairo/src/set-access-control.ts
@@ -1,3 +1,4 @@
+import { withImplicitArgs } from './common-options';
 import type { ContractBuilder, BaseFunction } from './contract';
 import { defineFunctions } from './utils/define-functions';
 import { defineModules } from './utils/define-modules';
@@ -9,12 +10,13 @@ export type Access = typeof accessOptions[number];
 export function setAccessControl(c: ContractBuilder, fn: BaseFunction, access: Access) {
   switch (access) {
     case 'ownable': {
-      c.addModule(modules.Ownable, [{ lit:'owner' }]);
-      c.addModuleFunction(
-        modules.Ownable, 'Ownable_initializer'
-      );
+      c.addModule(modules.Ownable, [{ lit:'owner' }], [], true);
       c.addConstructorArgument({ name: 'owner', type: 'felt'});
-      c.addLibraryCall(functions.only_owner, fn);
+      c.addLibraryCall(functions.assert_only_owner, fn);
+
+      c.addFunction(functions.transferOwnership);
+      c.addFunction(functions.renounceOwnership);
+
       break;
     }
   }
@@ -23,14 +25,34 @@ export function setAccessControl(c: ContractBuilder, fn: BaseFunction, access: A
 const modules = defineModules( {
   Ownable: {
     path: 'openzeppelin/access/ownable',
-    usePrefix: true
+    useNamespace: true
   },
 })
 
 const functions = defineFunctions({
   // --- library-only calls ---
-  only_owner: {
+  assert_only_owner: {
     module: modules.Ownable,
     args: [],
+  },
+
+  // --- external functions ---
+
+  transferOwnership: {
+    module: modules.Ownable,
+    kind: 'external' as const,
+    implicitArgs: withImplicitArgs(),
+    args: [
+      { name: 'newOwner', type: 'felt' },
+    ],
+    parentFunctionName: 'transfer_ownership',
+  },
+
+  renounceOwnership: {
+    module: modules.Ownable,
+    kind: 'external' as const,
+    implicitArgs: withImplicitArgs(),
+    args: [],
+    parentFunctionName: 'renounce_ownership',
   },
 });