+++ packages/core-cairo/src/utils/module-prefix.ts
@@ -1,12 +1,27 @@
-import type { BaseFunction, Module } from "../contract";
+import type { BaseFunction } from "../contract";
 
-function getPrefix(module: Module | undefined): string {
-  if (module === undefined) {
-    return '';
+/**
+ * If the function's module has a namespace, returns the namespace.
+ * Otherwise returns the function name itself.
+ */
+export function getImportName(fn: BaseFunction): string {
+  if (fn.module?.useNamespace) {
+    return fn.module.name;
+  } else {
+    return getFunctionName(fn);
   }
-  return module.usePrefix ? `${module.name}_` : '';
 }
 
+/**
+ * Returns the function name with either namespace or module prefix based on extensibility pattern. 
+ */
 export function getFunctionName(fn: BaseFunction): string {
-  return `${getPrefix(fn.module)}${fn.parentFunctionName ?? fn.name}`;
+  const suffix = fn.parentFunctionName ?? fn.name;
+  let prefix: string;
+  if (fn.module !== undefined && fn.module.useNamespace) {
+    prefix = `${fn.module.name}.`
+  } else {
+    prefix = '';
+  }
+  return `${prefix}${suffix}`;
 }